requirements.txt
black==26.1.0
certifi==2026.2.25
charset-normalizer==3.4.5
click==8.3.1
greenlet==3.3.2
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mypy_extensions==1.1.0
packaging==26.0
pathspec==1.0.4
platformdirs==4.9.4
playwright==1.58.0
pluggy==1.6.0
pyee==13.0.1
Pygments==2.19.2
pytest==9.0.2
pytest-base-url==2.1.0
pytest-html==4.2.0
pytest-metadata==3.1.1
pytest-playwright==0.7.2
python-slugify==8.0.4
pytokens==0.4.1
requests==2.32.5
text-unidecode==1.3
typing_extensions==4.15.0
urllib3==2.6.3
psycopg2-binary==2.9.11
python-dotenv==1.2.2
pytest-xdist==3.6.1
